Move FunctionIOManager into a separate module
